test salarieCDI + modifs class

+++ src/test/java/exercices/TP2/SalaireTest.java
@@ -12,11 +12,11 @@ class SalaireTest {
     @CsvSource({"10"})
     public void payer( int tauxHoraire){
         Salaire salaire = new Salaire(tauxHoraire);
-        int salaireNormal = salaire.payer(140); // 140 heures travaillées
+        double salaireNormal = salaire.payer(140); // 140 heures travaillées
         assertEquals(1400, salaireNormal); // 140 heures * 10 (taux horaire) = 1400 euros
 
         // Test pour les heures supplémentaires
-        int salaireSupplementaire = salaire.payer(160); // 160 heures travaillées
+        double salaireSupplementaire = salaire.payer(160); // 160 heures travaillées
         assertEquals(1622, salaireSupplementaire); // 151.67 heures * 10 + 25% * (160 - 151.67) * 10 = 1622 euros
     }
